requirements.txt
mlx-audio>=0.4.0
sounddevice
soundfile
pyobjc-framework-Quartz
pyobjc-framework-Cocoa
